Update boringssl and curl(v7.50.0)

+++ packager/media/base/test/fake_prng.cc
@@ -19,7 +19,7 @@ FILE* g_rand_source_fp = NULL;
 const char kFakePrngDataFile[] = "fake_prng_data.bin";
 
 // RAND_bytes and RAND_pseudorand implementation.
-int FakeBytes(uint8_t* buf, int num) {
+int FakeBytes(uint8_t* buf, size_t num) {
   DCHECK(buf);
   DCHECK(g_rand_source_fp);
